pyproject.toml
[project]
name = "repo-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cocoindex>=0.1.82",
    "cocoindex[embeddings]>=0.1.82",
    "numpy>=2.3.2",
    "pgvector>=0.4.1",
    "psycopg[binary,pool]>=3.2.9",
    "python-dotenv>=1.0.1",
]
